use the correct calling convention
the expected order of calling is that #on should be called first (to register)
the event blocks) and then for every time #utter is called with a particular
event, the associated registered block should be called.

the previous code didn't correctly do this order (since i was still platying
around with the syntax) and this commit fixes that.